// spadata/spa-data/ngx-data-requests/src/shared/interfaces/dataRequestdto.ts
export class DataRequestDTO {
    accountingDate: string;
    checkAll: string;
    code: string;
    creator: string;
    criteriaInfo: string;
    data: string;
    dateFrom: string;
    dateTo: string;
    documentChannel: string;
    documentDescription: string;
    documentStorageType: string;
    documentType: string;
    email: string;
    extractionType: string;
    format: string;
    policyTypeGroup: string;
    product: string;
    receipt: string;
    register: string;
    selectionTypeDate: string;
    subAgentNumber: string;
    trimester: string;
    year: string;
}